Move App into its own folder in src/components
