+++ src/components/Nav/Nav.css
@@ -2,3 +2,8 @@
   display: flex;
   justify-content: space-around;
 }
+
+.logo {
+  height: 20px;
+  /* width: 10%; */
+}
